Create lorem-standalone.js
I've slightly modified the script for people like me who want to use this library as a way dynamically insert Lorem Ipsum into their JavaScript/web app's interface/mock.

I also added a new "TITLE" type to generate titles (for my own and other's needs) :)

This is meant to be contributed as a separate file for the purposes mentioned above. I'd also be happy to write up a read-me for this!